requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
requests==2.28.1
python-dotenv==1.0.0
cachetools==5.3.1